+++ cmd/fireantelope/cli/firehose.go
