+++ cmd/fireantelope/cli/tools.go
@@ -14,7 +14,7 @@
 
 package cli
 
-import "github.com/streamingfast/firehose-acme/tools"
+import "github.com/EOS-Nation/firehose-antelope/tools"
 
 func init() {
 	RootCmd.AddCommand(tools.Cmd)
