+++ codec/console_reader_test.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/streamingfast/firehose-acme/types"
+	"github.com/EOS-Nation/firehose-antelope/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
